Applied review comments   * Havent moved yet to non snapshot testing   * throw validation error rather than type error   * use loadash for array flattening
