+++ src/lib/utils/validation.ts
@@ -1,3 +1,4 @@
+import * as tf from '@tensorflow/tfjs';
 import * as _ from 'lodash';
 import { isArray } from 'util';
 import { Type1DMatrix, Type2DMatrix, TypeMatrix } from '../types';
@@ -164,3 +165,35 @@ export const validateFeaturesConsistency = <T>(
     );
   }
 };
+
+/**
+ * Checks that provided X matrix has the same number of features as model matrix
+ * @param y_true - matrix to check
+ * @param y_pred - matrix to check
+ * @throws ValidationError - in case any of the params are empty
+ * @throws ValidationError - in case y_true and y_pred are of different shape
+ * @ignore
+ */
+export function validateShapesEqual(
+  y_true: Type1DMatrix<number> | Type2DMatrix<number> = null,
+  y_pred: Type1DMatrix<number> | Type2DMatrix<number> = null,
+): tf.Tensor[] {
+  const yTrueTensor = tf.tensor(y_true);
+  const yPredTensor = tf.tensor(y_pred);
+  const yTrueShape = inferShape(y_true);
+  const yPredShape = inferShape(y_pred);
+
+  // Validation 1: empty array check
+  if (yTrueShape[0] === 0 || yPredShape[0] === 0) {
+    throw new ValidationError(`y_true ${JSON.stringify(y_true)} and y_pred ${JSON.stringify(y_pred)} cannot be empty`);
+  }
+
+  // Validation 2: Same shape
+  if (!_.isEqual(yTrueShape, yPredShape)) {
+    throw new ValidationError(
+      `Shapes of y_true ${JSON.stringify(yTrueShape)} and y_pred ${JSON.stringify(yPredShape)} should be equal`,
+    );
+  }
+
+  return [yTrueTensor, yPredTensor];
+}
